pyproject.toml
[project]
name = "fedpoffa-ai-dept-curriculum-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.8",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "langchain-text-splitters>=0.3.8",
    "openai>=1.78.0",
    "pinecone>=7.0.0",
    "pydantic>=2.11.4",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
]
